+++ VTK/Utilities/vtkexodus2/expvan.c
@@ -55,7 +55,7 @@
 *
 * revision history - 
 *
-*  $Id: expvan.c,v 1.2 2006-11-29 18:09:13 dcthomp Exp $
+*  $Id: expvan.c,v 1.3 2008-04-22 21:14:53 david.cole Exp $
 *
 *****************************************************************************/
 
@@ -174,7 +174,7 @@ int ex_put_var_names (int   exoid,
      start[1] = 0;
 
      count[0] = 1;
-     count[1] = strlen(var_names[i]) + 1;
+     count[1] = (long)strlen(var_names[i]) + 1;
 
      if (ncvarput (exoid, varid, start, count, (void*) var_names[i]) == -1)
      {
